Move ReportController to rest, reorganise AdminService
