+++ src/main/java/com/imjustdoom/pluginsite/controller/FileController.java
@@ -20,7 +20,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Optional;
 
-@Controller
 @AllArgsConstructor
 public class FileController {
