+++ src/main/java/com/imjustdoom/pluginsite/controller/ProfileController.java
@@ -21,7 +21,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-@Controller
 @AllArgsConstructor
 public class ProfileController {
