+++ src/main/java/com/imjustdoom/pluginsite/controller/rest/AdminController.java
@@ -3,10 +3,12 @@ package com.imjustdoom.pluginsite.controller.rest;
 import com.imjustdoom.pluginsite.config.exception.RestErrorCode;
 import com.imjustdoom.pluginsite.config.exception.RestException;
 import com.imjustdoom.pluginsite.model.Report;
-import com.imjustdoom.pluginsite.service.rest.AdminService;
+import com.imjustdoom.pluginsite.service.rest.ReportService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,22 +20,22 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/admin")
 @RequiredArgsConstructor
 public class AdminController {
-    private final AdminService adminService;
+    private final ReportService reportService;
 
     @GetMapping("/report")
-    public Page<Report> listReports(Pageable pageable) throws RestException {
+    public Page<Report> listReports(@PageableDefault(sort = "reportedDate", direction = Sort.Direction.DESC) Pageable pageable) throws RestException {
         if (pageable.getPageSize() > 50) throw new RestException(RestErrorCode.PAGE_SIZE_TOO_LARGE, "Page size too large (%s > %s)", pageable.getPageSize(), 50);
-        return this.adminService.getReportPage(pageable);
+        return this.reportService.getReportPage(pageable);
     }
 
     @GetMapping("/report/{id}")
     public Report getReport(@PathVariable int id) throws RestException {
-        return this.adminService.getReport(id);
+        return this.reportService.getReport(id);
     }
 
     @PatchMapping("/report/{id}")
     public Report updateActionTaken(@PathVariable int id, @RequestParam String actionTaken) {
-        return this.adminService.updateActionTaken(id, actionTaken);
+        return this.reportService.updateActionTaken(id, actionTaken);
     }
 
     // todo didnt add roles since this system should be reworked. The code below is the old code.
